requirement.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.1.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.1
